+++ search/originals.cpp
@@ -151,7 +151,7 @@ string search_originals (void * webserver_request)
       v_passages.push_back (passage);
       counts.push_back (count);
     }
-    quick_sort (counts, v_passages, 0, counts.size());
+    quick_sort (counts, v_passages, 0, static_cast<int>(counts.size()));
     reverse (v_passages.begin(), v_passages.end());
